Disabled WebAssemblyThreads feature.
This change is due to failure of browser test
SharedArrayBufferDisabledTest.IsDisabled.

Needs security review.

In C73 this feature was disabled by default. In C74 it is enabled
by default on all platforms but Android. When enabled, this feature
also enables SharedArrayBuffer feature, which we have disabled due
to vulnerability to Spectre. See
https://github.com/brave/brave-browser/issues/1051.

Chromium change:

commit f9cc6b912e601070728b449892097cec0d59cbfa
Author: Ben Smith <binji@chromium.org>
Date:   Wed Feb 27 02:09:43 2019 +0000

    [wasm] Enable wasm threads by default

    See intent to ship blink-dev thread:
    https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/tD6np-OG2PU

    Bug: chromium:754910